% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioClimData.R
\name{RasterLayerCoordinates}
\alias{RasterLayerCoordinates}
\title{Raster Layer Coordinates}
\usage{
RasterLayerCoordinates(longitude, latitude, raster_layer)
}
\arguments{
\item{longitude}{A number -180 to 180}

\item{latitude}{A number -90 to 90}

\item{raster_layer}{A raster layer}
}
\value{
This function returns the values of a raster layer at the specified coordinates.
}
\description{
Raster Layer Coordinates
}

